% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispdm_report.R
\name{crispdm_report}
\alias{crispdm_report}
\title{CRISP-DM Report}
\usage{
crispdm_report(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \link[rmarkdown:render]{render}
}
\description{
Format for creating a CRISP-DM report. Adapted from \link{https://www.the-modeling-agency.com/crisp-dm.pdf}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyReport.Rmd", template = "crispdm_report", package = "rmddm")
}

}
