% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispdm_report.R
\name{crispdm_report}
\alias{crispdm_report}
\title{CRISP-DM Report}
\usage{
crispdm_report(file, edit = FALSE, create_dir = FALSE)
}
\arguments{
\item{file}{File name for the draft}

\item{edit}{\code{TRUE} to edit the template immediately}

\item{create_dir}{\code{TRUE} to create a new directory for the document
(the "default" setting leaves this behavior up to the creator of the
template).}
}
\value{
R Markdown output format to pass to \link[rmarkdown:render]{render}
}
\description{
Format for creating a CRISP-DM report. Adapted from \url{https://www.the-modeling-agency.com/crisp-dm.pdf}.
}
\examples{

\dontrun{
library(rmarkdown)
crispdm_report(name = "MyReport.Rmd",edit=TRUE,create_dir = FALSE)
}

}
\seealso{
\link[rmarkdown:draft]{draft}, \link[rmarkdown:render]{render}
}
